+++ src/meltano/core/utils/__init__.py
@@ -366,7 +366,23 @@ def set_at_path(d, path, value):
     final[tail] = value
 
 
-def expand_env_vars(raw_value, env: Dict):
+class EnvironmentVariableNotSetError(Exception):
+    """Occurs when a referenced environment variable is not set."""
+
+    def __init__(self, env_var: str):
+        """Initialize the error.
+        Args:
+            env_var: the unset environment variable name
+        """
+        super().__init__(env_var)
+        self.env_var = env_var
+
+    def __str__(self) -> str:
+        """Return the error as a string."""
+        return f"{self.env_var} referenced but not set."
+
+
+def expand_env_vars(raw_value, env: Dict, raise_if_missing: bool = False):
     if not isinstance(raw_value, str):
         return raw_value
 
@@ -394,8 +410,11 @@ def expand_env_vars(raw_value, env: Dict):
 
             return val
         except KeyError as e:
-            logger.debug(f"Variable '${var}' is missing from the environment.")
-            return None
+            if raise_if_missing:
+                raise EnvironmentVariableNotSetError(e.args[0])
+            else:
+                logger.debug(f"Variable '${var}' is missing from the environment.")
+                return None
 
     fullmatch = re.fullmatch(var_matcher, raw_value)
     if fullmatch: